refactor: move setting components into one folder
